エラーハンドリングを追加し、階層の整理を実施
