+++ lib/features/weather/screen/weather_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_training/features/weather/components/error_dialog.dart';
 import 'package:flutter_training/features/weather/components/weather.dart';
+import 'package:flutter_training/features/weather/exceptions/app_exception.dart';
 import 'package:flutter_training/features/weather/model/weather_condition.dart';
 import 'package:yumemi_weather/yumemi_weather.dart';
 
@@ -9,15 +10,6 @@ class WeatherScreen extends StatefulWidget {
 
   @override
   State<WeatherScreen> createState() => _WeatherScreenState();
-}
-
-extension YumemiWeatherErrorExt on YumemiWeatherError {
-  String convertErrorMessage() {
-    return switch (this) {
-      YumemiWeatherError.unknown => 'APIのエラーが発生しました。',
-      YumemiWeatherError.invalidParameter => '不正な値が入力されました。'
-    };
-  }
 }
 
 class _WeatherScreenState extends State<WeatherScreen> {
@@ -39,6 +31,14 @@ class _WeatherScreenState extends State<WeatherScreen> {
         barrierDismissible: false,
         builder: (_) {
           return ErrorDialogWidget(errorMessage: e.convertErrorMessage());
+        },
+      );
+    } on WeatherConditionException catch (e) {
+      await showDialog<void>(
+        context: context,
+        barrierDismissible: false,
+        builder: (_) {
+          return ErrorDialogWidget(errorMessage: e.toString());
         },
       );
     }
